rename logFactory.js to makeLog.js and move to utils
